refactor(tools-builtin): ADR-156 §6.4 mod.rs slim down
Per ADR-156 §6.4, after F4.6.1 ~ F4.6.8 sink waves the desktop
`tools::builtin` aggregator is now a thin re-export entry: each sunk
`dasclaw_*_tools` crate is wildcard re-exported so every historical
`crate::tools::builtin::<Tool>` call site keeps compiling unchanged.

Changes:
- desktop-client/ironclaw/src/tools/builtin/mod.rs: replace explicit
  per-crate type listings with `pub use dasclaw_*_tools::*;` wildcards;
  drop `pub mod shell` and `pub mod lsp` (their content is fully
  in dasclaw_shell_tools / dasclaw_lsp); keep `pub mod` declarations
  only for T3/T4 desktop-resident families (memory, tool_info,
  extension_tools, skill_tools, job, routine, message — see
  ADR-156 §6.3.1 + §8 / ADR-155).
- desktop-client/ironclaw/src/tools/builtin/shell.rs: delete shim.
- desktop-client/ironclaw/src/tools/builtin/lsp/mod.rs: delete shim.
- crates/dasclaw_fs_tools/src/lib.rs: add top-level type re-exports
  (CodeEditTool / GlobSearchTool / GrepSearchTool /
  ApplyPatchTool / ListDirTool / ReadFileTool / WriteFileTool) so
  wildcard re-export at desktop side surfaces tool types
  consistently with sibling crates.
- Migrate 3 call sites:
  * desktop-client/ironclaw/src/agent/dispatcher.rs:
    `tools::builtin::shell::classify_command_risk` →
    `tools::builtin::classify_command_risk`.
  * desktop-client/ironclaw/src/tools/registry.rs:
    `tools::builtin::lsp::LspRegistry` →
    `tools::builtin::LspRegistry`.
  * desktop-client/ironclaw/tests/parity_gate_p1.rs: same.

Verification (local):
- cargo check -p dasclaw --tests  — clean
- cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings — clean
- cargo nextest run -p dasclaw -E 'test(parity_gate_p1) | test(dispatcher) | test(registry)' — 204/204 pass
- cargo fmt --all  — clean
- python3.12 scripts/check_no_panics.py --base origin/xClaw — OK
- python3.12 scripts/check_no_new_ironclaw_literal.py --base origin/xClaw — OK
- python3.12 scripts/check_blueprint_sync.py — OK (47 crates / 2 planned)

Cross-cuts: Class A (no .ironclaw / IRONCLAW_BASE_DIR literals).
